requirements.txt
# Core dependencies
flask==3.1.2
python-dotenv==1.1.1

# Database and ORM
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5

# Authentication and Security
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
flask-limiter==3.5.0

# Caching
flask-caching==2.1.0
redis==5.0.1

# Production WSGI server
gunicorn==21.2.0

# Optional but useful (recommended for future enhancements)
flask-cors==5.0.0       # For enabling frontend-backend communication
sqlalchemy==2.0.36       # ORM for database operations (optional but powerful)
werkzeug==3.1.3          # Flask's core dependency, ensures version consistency
jinja2==3.1.6            # Template rendering engine
itsdangerous==2.2.0      # Security for session cookies
blinker==1.9.0           # For Flask signals and extensions
click==8.3.0             # Command-line interface for Flask commands

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
